[#18] Move methods to parent class
